cli: move sopel.run() into sopel.cli.run.run()
